refactor: unify type restoration via restore_dataclass_types
Move datetime/bytes field restoration from ad-hoc per-class methods
into a single restore_dataclass_types() in serialization.py. It
introspects dataclass annotations and uses match/case to restore
fields — no hardcoded field name lists needed.

InvocationResponse.from_json and Result.load now both delegate to it,
removing str_to_datetime from their imports and eliminating the
_parse_datetime_fields / _get_type_args helpers from results.py.
